+++ flask_typed_routes/__init__.py
@@ -2,7 +2,7 @@
 
 from flask_typed_routes.app import FlaskTypedRoutes, Mode, typed_route  # noqa
 from flask_typed_routes.errors import ValidationError  # noqa
-from flask_typed_routes.fields import Query, Path, Header, Cookie, Body  # noqa
+from flask_typed_routes.fields import Query, Path, Header, Cookie, Body, Depends  # noqa
 
 __all__ = [
     "ValidationError",
@@ -11,6 +11,7 @@ __all__ = [
     "Header",
     "Cookie",
     "Body",
+    "Depends",
     "typed_route",
     "FlaskTypedRoutes",
     "Mode",