+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Releases
 
+- **Add**: Add `Depends` field to allow dependencies between parameters.
+- **Add**: The `typed_route` decorator can receive `dependencies`.
+- **Change**: **OpenApi** schema has been improved for complex serialization scenarios like JSON in query params.
+
 ### v0.2.4 (2025-02-07)
 
 - **Add**: Refactor to improve code organization